EvolveFuzzySeperator_r: float-typed rand-const at gating sites
PTR ds:0x10058218 (= 0.000030518509f). Constants 0.01 and 0.5

Source-commit: 44af192459421e8a3e9fd9df1ff691de0a4eda83

+++ botlib/botlib.c
@@ -32332,10 +32332,10 @@ void __cdecl EvolveFuzzySeperator_r(fuzzyseperator_t *fs)
     }
     else if ( fs->type == 1 )
     {
-      if ( (double)(rand() & 0x7FFF) * 0.000030518509 >= 0.01 )
+      if ( (float)(rand() & 0x7FFF) * 0.000030518509f >= 0.01 )
       {
         v4 = rand();
-        v3 = ((double)(v4 & 0x7FFF) * 0.000030518509 - 0.5 + (double)(v4 & 0x7FFF) * 0.000030518509 - 0.5)
+        v3 = ((float)(v4 & 0x7FFF) * 0.000030518509f - 0.5 + (float)(v4 & 0x7FFF) * 0.000030518509f - 0.5)
            * (fs->maxweight - fs->minweight)
            * 0.5;
       }
